Moved Extended Euclidean Algorithm from risch.py to densetools.py
Moved Extended Euclidean Algorithm, Diophantine version that is gcdex_diophantine from risch.py in integral module to densetools.poly in polys module
