+++ sympy/integrals/risch.py
@@ -36,7 +36,7 @@ from sympy.core.compatibility import reduce, ordered
 from sympy.functions import log, exp, sin, cos, tan, asin, acos, atan
 
 from sympy.integrals import Integral, integrate
-
+from sympy.polys.densetools import gcdex_diophantine
 from sympy.polys import gcd, cancel, PolynomialError, Poly, reduced, RootSum, DomainError
 
 from sympy.utilities.iterables import numbered_symbols
@@ -669,31 +669,6 @@ class NonElementaryIntegralException(Exception):
     # TODO: Pass through information about why the integral was nonelementary,
     # and store that in the resulting NonElementaryIntegral somehow.
     pass
-
-
-def gcdex_diophantine(a, b, c):
-    """
-    Extended Euclidean Algorithm, Diophantine version.
-
-    Given a, b in K[x] and c in (a, b), the ideal generated by a and b,
-    return (s, t) such that s*a + t*b == c and either s == 0 or s.degree()
-    < b.degree().
-    """
-    # Extended Euclidean Algorithm (Diophantine Version) pg. 13
-    # TODO: This should go in densetools.py.
-    # XXX: Bettter name?
-
-    s, g = a.half_gcdex(b)
-    q = c.exquo(g)  # Inexact division means c is not in (a, b)
-    s = q*s
-
-    if not s.is_zero and b.degree() >= b.degree():
-        q, s = s.div(b)
-
-    t = (c - s*a).exquo(b)
-
-    return (s, t)
-
 
 def frac_in(f, t, **kwargs):
     """
