+++ sympy/integrals/tests/test_risch.py
@@ -2,7 +2,10 @@
 from __future__ import with_statement
 from sympy import (Poly, S, Function, log, symbols, exp, tan, sqrt,
     Symbol, Lambda, sin)
-from sympy.integrals.risch import (gcdex_diophantine, frac_in, as_poly_1t,
+from sympy import (Poly, I, S, Function, log, symbols, exp, tan, sqrt,
+    Symbol, Lambda, sin, cos)
+from sympy.polys.densetools import gcdex_diophantine
+from sympy.integrals.risch import (frac_in, as_poly_1t,
     derivation, splitfactor, splitfactor_sqf, canonical_representation,
     hermite_reduce, polynomial_reduce, residue_reduce, residue_reduce_to_basic,
     integrate_primitive, integrate_hyperexponential_polynomial,
